ai_studio/requirements.txt
aiohttp>=3.8.5
python-dotenv>=1.0.0
apscheduler>=3.10.1
selenium>=4.15.0
webdriver-manager>=4.0.0
openai>=1.3.0
anthropic>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.2
sqlite3
asyncio>=3.4.3
